% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{add_calculated_column}
\alias{add_calculated_column}
\title{Add calculated column}
\usage{
add_calculated_column(m, col_name, calculation_func, to_rows = TRUE)
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{col_name}{A character string specifying the name of the new column.}

\item{calculation_func}{A function that takes a list (row of index data.frame) and returns a value.}

\item{to_rows}{A logical value indicating whether to add the column to row indices (TRUE) or column indices (FALSE) (default: TRUE).}
}
\value{
A MatrixEntry object.
}
\description{
Add a calculated column to row or column indices based on existing index values.
}
\details{
Add a calculated column to row or column indices
}
\examples{
\dontrun{
me_new <- add_calculated_column(me, "is_developed", function(row) row$Country == "USA")
}
}
