% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate GroupedMatrixEntry}
\usage{
aggregate(gm, func)
}
\arguments{
\item{gm}{A GroupedMatrixEntry object.}

\item{func}{An R function (e.g. sum, mean) or character string naming a Julia function (e.g. "sum", "mean").}
}
\value{
A MatrixEntry object.
}
\description{
Aggregate matrix data grouped by groupby.
}
\details{
Aggregate a GroupedMatrixEntry using a function
}
\examples{
\dontrun{
aggregated <- aggregate(gm, "sum")
}
}
