% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{as.data.frame.MatrixEntry}
\alias{as.data.frame.MatrixEntry}
\title{Convert MatrixEntry to long DataFrame}
\usage{
\method{as.data.frame}{MatrixEntry}(x, row.names = NULL, optional = FALSE, ..., value_name = "value")
}
\arguments{
\item{x}{A MatrixEntry object.}

\item{row.names}{Unused.}

\item{optional}{Unused.}

\item{...}{Unused.}

\item{value_name}{A character string specifying the name of the value column (default: "value").}
}
\value{
A data.frame.
}
\description{
Convert a MatrixEntry to a long-form DataFrame.
}
\details{
Convert MatrixEntry to long DataFrame
}
