% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{as.data.frame.SeriesEntry}
\alias{as.data.frame.SeriesEntry}
\title{Convert SeriesEntry to DataFrame}
\usage{
\method{as.data.frame}{SeriesEntry}(x, row.names = NULL, optional = FALSE, ..., value_name = "value")
}
\arguments{
\item{x}{A SeriesEntry object.}

\item{row.names}{Unused.}

\item{optional}{Unused.}

\item{...}{Unused.}

\item{value_name}{A character string specifying the name of the value column (default: "value").}
}
\value{
A data.frame.
}
\description{
Convert a SeriesEntry to a DataFrame.
}
\details{
Convert SeriesEntry to DataFrame
}
