% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{countries}
\alias{countries}
\title{Get available countries}
\usage{
countries(x)
}
\arguments{
\item{x}{An MRIO, MatrixEntry, SeriesEntry, EnvironmentalExtension, or data.frame object.}
}
\value{
A character vector of unique country codes.
}
\description{
Retrieve a character vector of unique country codes/names available in the MRIO database or data structure.
}
\details{
Get unique countries in the MRIO database or data structure
}
\examples{
\dontrun{
countries(mrio_db)
}
}
