% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{country_summary}
\alias{country_summary}
\title{Country summary of MatrixEntry}
\usage{
country_summary(m)
}
\arguments{
\item{m}{A MatrixEntry object.}
}
\value{
A data.frame.
}
\description{
Generate country-by-country flow summary for bilateral analysis.
}
\details{
Generate country-by-country flow summary
}
\examples{
\dontrun{
bilateral_flows <- country_summary(me)
}
}
