% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{dim.LeontiefFactorization}
\alias{dim.LeontiefFactorization}
\title{Dimensions of LeontiefFactorization}
\usage{
\method{dim}{LeontiefFactorization}(x)
}
\arguments{
\item{x}{A LeontiefFactorization object.}
}
\value{
An integer vector of length 2.
}
\description{
Dimensions of LeontiefFactorization
}
