% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{dim.MatrixEntry}
\alias{dim.MatrixEntry}
\title{Dimensions of MatrixEntry}
\usage{
\method{dim}{MatrixEntry}(x)
}
\arguments{
\item{x}{A MatrixEntry object.}
}
\value{
An integer vector of length 2.
}
\description{
Dimensions of MatrixEntry
}
