% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.DollarNames.MRIO}
\alias{.DollarNames.MRIO}
\title{Autocomplete names for MRIO}
\usage{
\method{.DollarNames}{MRIO}(x, pattern = "")
}
\arguments{
\item{x}{An MRIO object.}

\item{pattern}{A character string to match.}
}
\value{
A character vector of matching field names.
}
\description{
Autocomplete names for MRIO
}
