% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.DollarNames.SeriesEntry}
\alias{.DollarNames.SeriesEntry}
\title{Autocomplete names for SeriesEntry}
\usage{
\method{.DollarNames}{SeriesEntry}(x, pattern = "")
}
\arguments{
\item{x}{A SeriesEntry object.}

\item{pattern}{A character string to match.}
}
\value{
A character vector of matching field names.
}
\description{
Autocomplete names for SeriesEntry
}
