% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{Eora}
\alias{Eora}
\title{Load Eora database}
\usage{
Eora(path)
}
\arguments{
\item{path}{A character string representing the directory path containing Eora database files.}
}
\value{
An MRIO object wrapping the Julia MRIO database.
}
\description{
Load and construct complete Eora MRIO database from a file directory.
}
\details{
Load and construct complete Eora MRIO database
}
\examples{
\dontrun{
eora_db <- Eora("data/2017/")
}
}
