% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{filter.MRIO}
\alias{filter.MRIO}
\title{Filter MRIO rows or columns}
\usage{
\method{filter}{MRIO}(.data, ..., .dims = 1)
}
\arguments{
\item{.data}{An MRIO object.}

\item{...}{Logical expressions.}

\item{.dims}{Dimension to filter: 1 for rows, 2 for columns.}
}
\value{
A filtered MRIO object.
}
\description{
Filter MRIO rows or columns
}
