% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{filter.SeriesEntry}
\alias{filter.SeriesEntry}
\title{Filter SeriesEntry}
\usage{
\method{filter}{SeriesEntry}(.data, ...)
}
\arguments{
\item{.data}{A SeriesEntry object.}

\item{...}{Logical expressions evaluated on col_indices.}
}
\value{
A SeriesEntry object.
}
\description{
Filter elements of a SeriesEntry using dplyr syntax.
}
\details{
Filter elements of a SeriesEntry using dplyr syntax
}
