% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{filter_cols}
\alias{filter_cols}
\title{Filter columns of MatrixEntry}
\usage{
filter_cols(m, condition_func)
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{condition_func}{A function that takes a list (column of index data.frame) and returns a logical value.}
}
\value{
A MatrixEntry object.
}
\description{
Filter columns based on a condition function applied to column indices.
}
\details{
Filter columns of a MatrixEntry based on a condition function
}
\examples{
\dontrun{
filtered <- filter_cols(me, function(col) col$Country == "CHN")
}
}
