% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter both dimensions of MatrixEntry}
\usage{
filter_matrix(m, row_condition, col_condition)
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{row_condition}{A function that takes a list (row of row_indices) and returns a logical.}

\item{col_condition}{A function that takes a list (column of col_indices) and returns a logical.}
}
\value{
A MatrixEntry object.
}
\description{
Filter both rows and columns using separate condition functions.
}
\details{
Subset a MatrixEntry by condition functions
}
\examples{
\dontrun{
filtered <- filter_matrix(me, function(r) r$Country == "USA", function(c) c$Sector == "Agr")
}
}
