% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{filter_rows}
\alias{filter_rows}
\title{Filter rows of MatrixEntry}
\usage{
filter_rows(m, condition_func)
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{condition_func}{A function that takes a list (row of index data.frame) and returns a logical value.}
}
\value{
A MatrixEntry object.
}
\description{
Filter rows based on a condition function applied to row indices.
}
\details{
Filter rows of a MatrixEntry based on a condition function
}
\examples{
\dontrun{
filtered <- filter_rows(me, function(row) row$Country == "USA")
}
}
