% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{find_julia_project}
\alias{find_julia_project}
\title{Find the directory containing Project.toml}
\usage{
find_julia_project()
}
\value{
A character string representing the directory path.
}
\description{
Find the directory containing Project.toml
}
