% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{from_long_dataframe}
\alias{from_long_dataframe}
\title{Long DataFrame to MatrixEntry}
\usage{
from_long_dataframe(
  df,
  value_col = "value",
  row_prefix = "row_",
  col_prefix = "col_"
)
}
\arguments{
\item{df}{A data.frame.}

\item{value_col}{A character string specifying the name of the value column (default: "value").}

\item{row_prefix}{A character string specifying the row columns prefix (default: "row_").}

\item{col_prefix}{A character string specifying the column columns prefix (default: "col_").}
}
\value{
A MatrixEntry object.
}
\description{
Convert a long-form DataFrame back to MatrixEntry format.
}
\details{
Convert a long-form DataFrame back to MatrixEntry format
}
\examples{
\dontrun{
me <- from_long_dataframe(df)
}
}
