% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_julia_connection}
\alias{get_julia_connection}
\title{Get or initialize the Julia connection and load the Juliora package}
\usage{
get_julia_connection()
}
\value{
Invisible NULL.
}
\description{
Get or initialize the Julia connection and load the Juliora package
}
