% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{Gloria}
\alias{Gloria}
\title{Load Gloria database}
\usage{
Gloria(path, version, year)
}
\arguments{
\item{path}{A character string representing the directory path containing Gloria database files.}

\item{version}{An integer specifying the Gloria version (e.g. 60).}

\item{year}{An integer specifying the database year.}
}
\value{
An MRIO object wrapping the Julia MRIO database.
}
\description{
Load and construct complete Gloria MRIO database from a file directory.
}
\details{
Load and construct complete Gloria MRIO database
}
\examples{
\dontrun{
gloria_db <- Gloria("data/GLORIA/2019/", 60, 2019)
}
}
