% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{is_julia_available}
\alias{is_julia_available}
\title{Check Julia availability}
\usage{
is_julia_available()
}
\value{
A logical value indicating whether Julia is available.
}
\description{
Verifies if a working Julia installation can be found by JuliaConnectoR.
}
\details{
Check if Julia environment is available
}
\examples{
\dontrun{
is_julia_available()
}
}
