% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{length.SeriesEntry}
\alias{length.SeriesEntry}
\title{Length of SeriesEntry}
\usage{
\method{length}{SeriesEntry}(x)
}
\arguments{
\item{x}{A SeriesEntry object.}
}
\value{
An integer.
}
\description{
Length of SeriesEntry
}
