% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{names.MatrixEntry}
\alias{names.MatrixEntry}
\title{Get names of MatrixEntry fields}
\usage{
\method{names}{MatrixEntry}(x)
}
\arguments{
\item{x}{A MatrixEntry object.}
}
\value{
A character vector of field names.
}
\description{
Get names of MatrixEntry fields
}
