% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{names.MRIO}
\alias{names.MRIO}
\title{Get names of MRIO fields}
\usage{
\method{names}{MRIO}(x)
}
\arguments{
\item{x}{An MRIO object.}
}
\value{
A character vector of field names.
}
\description{
Get names of MRIO fields
}
