% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{names.SeriesEntry}
\alias{names.SeriesEntry}
\title{Get names of SeriesEntry fields}
\usage{
\method{names}{SeriesEntry}(x)
}
\arguments{
\item{x}{A SeriesEntry object.}
}
\value{
A character vector of field names.
}
\description{
Get names of SeriesEntry fields
}
