% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bindings.R
\name{new_community_result}
\alias{new_community_result}
\title{Create a CommunityResult R object}
\usage{
new_community_result(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy to wrap.}
}
\value{
A CommunityResult S3 object.
}
\description{
Create a CommunityResult R object
}
