% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{new_environmental_extension}
\alias{new_environmental_extension}
\title{Create an EnvironmentalExtension R object}
\usage{
new_environmental_extension(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy to wrap.}
}
\value{
An EnvironmentalExtension S3 object.
}
\description{
Create an EnvironmentalExtension R object
}
