% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{new_leontief_factorization}
\alias{new_leontief_factorization}
\title{Create a LeontiefFactorization R object}
\usage{
new_leontief_factorization(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy to wrap.}
}
\value{
A LeontiefFactorization S3 object.
}
\description{
Create a LeontiefFactorization R object
}
