% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{new_matrix_entry}
\alias{new_matrix_entry}
\title{Create a MatrixEntry R object}
\usage{
new_matrix_entry(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy to wrap.}
}
\value{
A MatrixEntry S3 object.
}
\description{
Create a MatrixEntry R object
}
