% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bindings.R
\name{new_mrio_graph}
\alias{new_mrio_graph}
\title{Create an MRIOGraph R object}
\usage{
new_mrio_graph(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy to wrap.}
}
\value{
An MRIOGraph S3 object.
}
\description{
Create an MRIOGraph R object
}
