% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{new_series_entry}
\alias{new_series_entry}
\title{Create a SeriesEntry R object}
\usage{
new_series_entry(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy to wrap.}
}
\value{
A SeriesEntry S3 object.
}
\description{
Create a SeriesEntry R object
}
