% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{parse_gloria}
\alias{parse_gloria}
\title{Parse Gloria database}
\usage{
parse_gloria(path, year, version = 60)
}
\arguments{
\item{path}{A character string representing the directory path containing Gloria database files.}

\item{year}{An integer specifying the database year.}

\item{version}{An integer specifying the Gloria version (default: 60).}
}
\value{
An MRIO object wrapping the Julia MRIO database.
}
\description{
Parse Gloria database files for a given year.
}
\details{
Parse Gloria database files
}
\examples{
\dontrun{
db <- parse_gloria("data/GLORIA/", 2019, version = 60)
}
}
