% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{pivot_matrix_to_wide}
\alias{pivot_matrix_to_wide}
\title{Pivot MatrixEntry to wide}
\usage{
pivot_matrix_to_wide(m, row_vars, col_var, value_var = "value")
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{row_vars}{A character vector of row variables to keep.}

\item{col_var}{A character string specifying the column variable to pivot.}

\item{value_var}{A character string specifying the value column (default: "value").}
}
\value{
A data.frame.
}
\description{
Pivot the matrix data to wide format for analysis or visualization.
}
\details{
Pivot the matrix data to wide format
}
\examples{
\dontrun{
df_wide <- pivot_matrix_to_wide(me, c("Country"), "Sector")
}
}
