% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{relocate.SeriesEntry}
\alias{relocate.SeriesEntry}
\title{Relocate SeriesEntry metadata columns}
\usage{
\method{relocate}{SeriesEntry}(.data, ...)
}
\arguments{
\item{.data}{A SeriesEntry object.}

\item{...}{Relocate expressions.}
}
\value{
A SeriesEntry object.
}
\description{
Relocate SeriesEntry metadata using dplyr syntax.
}
\details{
Relocate SeriesEntry metadata using dplyr syntax
}
