% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{rename.SeriesEntry}
\alias{rename.SeriesEntry}
\title{Rename SeriesEntry metadata columns}
\usage{
\method{rename}{SeriesEntry}(.data, ...)
}
\arguments{
\item{.data}{A SeriesEntry object.}

\item{...}{Rename expressions.}
}
\value{
A SeriesEntry object.
}
\description{
Rename SeriesEntry metadata using dplyr syntax.
}
\details{
Rename SeriesEntry metadata using dplyr syntax
}
