% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{slice.SeriesEntry}
\alias{slice.SeriesEntry}
\title{Slice SeriesEntry}
\usage{
\method{slice}{SeriesEntry}(.data, ...)
}
\arguments{
\item{.data}{A SeriesEntry object.}

\item{...}{Slicing expressions.}
}
\value{
A SeriesEntry object.
}
\description{
Slice elements of a SeriesEntry using dplyr syntax.
}
\details{
Slice elements of a SeriesEntry using dplyr syntax
}
