% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{stressor}
\alias{stressor}
\title{Get available stressors (alias)}
\usage{
stressor(x)
}
\arguments{
\item{x}{An MRIO, MatrixEntry, SeriesEntry, EnvironmentalExtension, or data.frame object.}
}
\value{
A character vector of unique stressor names.
}
\description{
Retrieve a character vector of unique stressor names available in the MRIO database or environmental extension.
}
\details{
Get unique stressors in the MRIO database or environmental extension (alias)
}
\examples{
\dontrun{
stressor(mrio_db)
}
}
