% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{stressors}
\alias{stressors}
\title{Get available stressors}
\usage{
stressors(x)
}
\arguments{
\item{x}{An MRIO, MatrixEntry, SeriesEntry, EnvironmentalExtension, or data.frame object.}
}
\value{
A character vector of unique stressor names.
}
\description{
Retrieve a character vector of unique stressor names available in the MRIO database or environmental extension.
}
\details{
Get unique stressors in the MRIO database or environmental extension
}
\examples{
\dontrun{
stressors(mrio_db)
}
}
