% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{[.MatrixEntry}
\alias{[.MatrixEntry}
\title{Subset MatrixEntry}
\usage{
\method{[}{MatrixEntry}(x, i, j, ...)
}
\arguments{
\item{x}{A MatrixEntry object.}

\item{i}{Rows to select (logical vector, named list, list of named lists, or missing).}

\item{j}{Columns to select (logical vector, named list, list of named lists, or missing).}

\item{...}{Unused.}
}
\value{
A single value, a SeriesEntry, or a MatrixEntry.
}
\description{
Subset MatrixEntry
}
