% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{[.SeriesEntry}
\alias{[.SeriesEntry}
\title{Subset SeriesEntry}
\usage{
\method{[}{SeriesEntry}(x, i, ...)
}
\arguments{
\item{x}{A SeriesEntry object.}

\item{i}{Column key (named list).}

\item{...}{Unused.}
}
\value{
A numeric value.
}
\description{
Subset SeriesEntry
}
