% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{sum_by_country}
\alias{sum_by_country}
\title{Sum MatrixEntry by country}
\usage{
sum_by_country(m, dimension = "both")
}
\arguments{
\item{m}{A MatrixEntry object}

\item{dimension}{A character string indicating the dimension to sum over: "both", "rows", or "cols" (default: "both")}
}
\value{
A data.frame containing the summed values by country
}
\description{
Sum matrix values by country codes.
}
\details{
Sum matrix values by country codes
}
\examples{
\dontrun{
df <- sum_by_country(me, dimension = "rows")
}
}
