% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{sum_by_sector}
\alias{sum_by_sector}
\title{Sum MatrixEntry by sector}
\usage{
sum_by_sector(m, dimension = "both")
}
\arguments{
\item{m}{A MatrixEntry object}

\item{dimension}{A character string indicating the dimension to sum over: "both", "rows", or "cols" (default: "both")}
}
\value{
A data.frame containing the summed values by sector
}
\description{
Sum matrix values by sector codes.
}
\details{
Sum matrix values by sector codes
}
\examples{
\dontrun{
df <- sum_by_sector(me, dimension = "cols")
}
}
