% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{summarise.GroupedMatrixEntry}
\alias{summarise.GroupedMatrixEntry}
\title{Summarize GroupedMatrixEntry}
\usage{
\method{summarise}{GroupedMatrixEntry}(.data, ..., .groups = NULL)
}
\arguments{
\item{.data}{A GroupedMatrixEntry object.}

\item{...}{Slicing expressions / aggregation functions (e.g. sum(value)).}

\item{.groups}{Unused.}
}
\value{
A MatrixEntry object.
}
\description{
Summarize and aggregate a GroupedMatrixEntry using dplyr syntax.
}
\details{
Summarize and aggregate a GroupedMatrixEntry using dplyr syntax
}
