% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{to_named_tuple}
\alias{to_named_tuple}
\title{Convert R named list to Julia NamedTuple proxy}
\usage{
to_named_tuple(x)
}
\arguments{
\item{x}{A named list.}
}
\value{
A Julia proxy object representing a NamedTuple.
}
\description{
Convert R named list to Julia NamedTuple proxy
}
