% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{to_named_tuple_vector}
\alias{to_named_tuple_vector}
\title{Convert R list of named lists to Julia Vector\{NamedTuple\} proxy}
\usage{
to_named_tuple_vector(x)
}
\arguments{
\item{x}{A list of named lists.}
}
\value{
A Julia proxy object representing a Vector of NamedTuples.
}
\description{
Convert R list of named lists to Julia Vector\{NamedTuple\} proxy
}
