% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{unwrap_julia_object}
\alias{unwrap_julia_object}
\title{Unwrap wrapped R S3 classes back to Julia proxies}
\usage{
unwrap_julia_object(x)
}
\arguments{
\item{x}{An object.}
}
\value{
The underlying JuliaProxy or the object itself.
}
\description{
Unwrap wrapped R S3 classes back to Julia proxies
}
