% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{wrap_julia_object}
\alias{wrap_julia_object}
\title{Wrap Julia proxies in R S3 classes}
\usage{
wrap_julia_object(proxy)
}
\arguments{
\item{proxy}{A JuliaProxy object.}
}
\value{
A wrapped object or the proxy itself.
}
\description{
Wrap Julia proxies in R S3 classes
}
