% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{group_by.MatrixEntry}
\alias{group_by.MatrixEntry}
\title{Group MatrixEntry}
\usage{
\method{group_by}{MatrixEntry}(.data, ..., .add = FALSE, .drop = dplyr::group_by_drop_default(.data))
}
\arguments{
\item{.data}{A MatrixEntry object.}

\item{...}{Columns to group by.}

\item{.add}{Unused.}

\item{.drop}{Unused.}
}
\value{
A GroupedMatrixEntry object.
}
\description{
Group MatrixEntry index metadata using dplyr syntax.
}
\details{
Group MatrixEntry index metadata using dplyr syntax
}
