% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{groupby_matrix}
\alias{groupby_matrix}
\title{Group and aggregate MatrixEntry}
\usage{
groupby_matrix(m, ..., agg_func = "sum", rows = TRUE, value_name = "value")
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{...}{Column names (symbols or character strings) to group by.}

\item{agg_func}{An R function or Julia function name (default: "sum" or sum).}

\item{rows}{A logical value indicating whether to group rows (TRUE) or columns (FALSE) (default: TRUE).}

\item{value_name}{A character string specifying the name of the value column (default: "value").}
}
\value{
A data.frame.
}
\description{
Group and aggregate matrix data by specified index columns.
}
\details{
Group and aggregate matrix data by specified index columns
}
\examples{
\dontrun{
df <- groupby_matrix(me, "CountryCode", agg_func = "sum")
}
}
