% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{induced_production}
\alias{induced_production}
\title{Leontief induced production}
\usage{
induced_production(mrio, consumer_countries, producer_countries)
}
\arguments{
\item{mrio}{An MRIO database object.}

\item{consumer_countries}{A character vector of consumer country codes.}

\item{producer_countries}{A character vector of producer country codes.}
}
\value{
A data.frame containing columns CountryCode, Sector, and InducedProduction.
}
\description{
Calculate the production induced by the final demand of specified consumer countries on specified producer countries.
}
\details{
Calculate induced production using the Leontief Inverse
}
\examples{
\dontrun{
res <- induced_production(mrio, c("DEU", "FRA"), c("IND", "BRA"))
}
}
