% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{mutate.MatrixEntry}
\alias{mutate.MatrixEntry}
\title{Mutate MatrixEntry metadata columns}
\usage{
\method{mutate}{MatrixEntry}(.data, ..., .dims = 1)
}
\arguments{
\item{.data}{A MatrixEntry object.}

\item{...}{Mutate expressions.}

\item{.dims}{An optional integer: 1 for row_indices, 2 for col_indices. Auto-detected by default.}
}
\value{
A MatrixEntry object.
}
\description{
Mutate MatrixEntry index metadata using dplyr syntax.
}
\details{
Mutate MatrixEntry index metadata using dplyr syntax
}
