% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R, R/dplyr_integrations.R
\name{mutate.SeriesEntry}
\alias{mutate.SeriesEntry}
\title{Mutate SeriesEntry metadata}
\usage{
\method{mutate}{SeriesEntry}(.data, ...)

\method{mutate}{SeriesEntry}(.data, ...)
}
\arguments{
\item{.data}{A SeriesEntry object.}

\item{...}{Mutate expressions.}
}
\value{
A mutated SeriesEntry.

A SeriesEntry object.
}
\description{
Mutate SeriesEntry metadata using dplyr syntax.
}
\details{
Mutate SeriesEntry metadata using dplyr syntax
}
