% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{parse_gloria_sut}
\alias{parse_gloria_sut}
\title{Parse Gloria SUT}
\usage{
parse_gloria_sut(path, year, version = 60)
}
\arguments{
\item{path}{A character string representing the directory path containing Gloria database files.}

\item{year}{An integer specifying the database year.}

\item{version}{An integer specifying the Gloria version (default: 60).}
}
\value{
An MRIO object wrapping the Julia MRIO database.
}
\description{
Parse Gloria Supply-Use Tables (SUT) files for a given year.
}
\details{
Parse Gloria Supply-Use Tables (SUT) files
}
\examples{
\dontrun{
sut_db <- parse_gloria_sut("data/GLORIA/", 2019, version = 60)
}
}
