% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{relocate.MatrixEntry}
\alias{relocate.MatrixEntry}
\title{Relocate MatrixEntry metadata columns}
\usage{
\method{relocate}{MatrixEntry}(.data, ..., .dims = 1)
}
\arguments{
\item{.data}{A MatrixEntry object.}

\item{...}{Relocate expressions.}

\item{.dims}{An optional integer: 1 for row_indices, 2 for col_indices. Auto-detected by default.}
}
\value{
A MatrixEntry object.
}
\description{
Relocate MatrixEntry index metadata using dplyr syntax.
}
\details{
Relocate MatrixEntry index metadata using dplyr syntax
}
