% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R, R/dplyr_integrations.R
\name{rename.MatrixEntry}
\alias{rename.MatrixEntry}
\title{Rename metadata columns in MatrixEntry}
\usage{
\method{rename}{MatrixEntry}(.data, ..., .dims = 1)

\method{rename}{MatrixEntry}(.data, ..., .dims = 1)
}
\arguments{
\item{.data}{A MatrixEntry object.}

\item{...}{Rename expressions.}

\item{.dims}{An optional integer: 1 for row_indices, 2 for col_indices. Auto-detected by default.}
}
\value{
A MatrixEntry object with renamed metadata columns.

A MatrixEntry object.
}
\description{
Rename MatrixEntry index metadata using dplyr syntax.
}
\details{
Rename MatrixEntry index metadata using dplyr syntax
}
