% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{sector}
\alias{sector}
\title{Get available sectors (alias)}
\usage{
sector(x)
}
\arguments{
\item{x}{An MRIO, MatrixEntry, SeriesEntry, EnvironmentalExtension, or data.frame object.}
}
\value{
A character vector of unique sector names.
}
\description{
Retrieve a character vector of unique sector names available in the MRIO database or data structure.
}
\details{
Get unique sectors in the MRIO database or data structure (alias)
}
\examples{
\dontrun{
sector(mrio_db)
}
}
