% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integrations.R
\name{select.MatrixEntry}
\alias{select.MatrixEntry}
\title{Select MatrixEntry metadata columns}
\usage{
\method{select}{MatrixEntry}(.data, ..., .dims = 1)
}
\arguments{
\item{.data}{A MatrixEntry object.}

\item{...}{Columns to select.}

\item{.dims}{An optional integer: 1 for row_indices, 2 for col_indices. Auto-detected by default.}
}
\value{
A MatrixEntry object.
}
\description{
Select columns from MatrixEntry index metadata using dplyr syntax.
}
\details{
Select columns from MatrixEntry index metadata using dplyr syntax
}
