% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R, R/dplyr_integrations.R
\name{select.SeriesEntry}
\alias{select.SeriesEntry}
\title{Select SeriesEntry metadata columns}
\usage{
\method{select}{SeriesEntry}(.data, ...)

\method{select}{SeriesEntry}(.data, ...)
}
\arguments{
\item{.data}{A SeriesEntry object.}

\item{...}{Columns to select.}
}
\value{
A SeriesEntry.

A SeriesEntry object.
}
\description{
Select columns from SeriesEntry metadata using dplyr syntax.
}
\details{
Select columns from SeriesEntry metadata using dplyr syntax
}
