% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{sum_rows}
\alias{sum_rows}
\title{Sum matrix rows}
\usage{
sum_rows(x)
}
\arguments{
\item{x}{A numeric matrix or MatrixEntry.}
}
\value{
A numeric vector with one value per row of \code{x}.
}
\description{
Calculate the sum of every row of a numeric matrix or MatrixEntry using the
Julia backend.
}
\details{
Sum matrix rows
}
\examples{
\dontrun{
sum_rows(matrix(1:6, nrow = 2))
}
}
