% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{to_long_dataframe}
\alias{to_long_dataframe}
\title{MatrixEntry to long DataFrame}
\usage{
to_long_dataframe(m, value_name = "value")
}
\arguments{
\item{m}{A MatrixEntry object.}

\item{value_name}{A character string specifying the name of the value column (default: "value").}
}
\value{
A data.frame.
}
\description{
Convert a MatrixEntry to long-form DataFrame suitable for R database/tidyverse operations.
}
\details{
Convert a MatrixEntry to long-form DataFrame
}
\examples{
\dontrun{
df <- to_long_dataframe(me)
}
}
